+++ src/SFManagement.Application/Commands/CreateWorkerCommand.cs
@@ -2,5 +2,5 @@ namespace SFManagement.Application.Commands;
 
 public record CreateWorkerCommand(string Name, string Role, float[] SkillsVector)
 {
-    public int CreatedId { get; set; }
+    public long CreatedId { get; set; }
 }
