+++ src/SFManagement.Application/Commands/RemoveWorkerFromTaskCommand.cs
@@ -1,3 +1,3 @@
 namespace SFManagement.Application.Commands;
 
-public record RemoveWorkerFromTaskCommand(int TaskId, int WorkerId);
+public record RemoveWorkerFromTaskCommand(long TaskId, long WorkerId);
