+++ src/SFManagement.Application/Commands/ToggleSkillActiveCommand.cs
@@ -1,3 +1,3 @@
 namespace SFManagement.Application.Commands;
 
-public record ToggleSkillActiveCommand(int SkillId);
+public record ToggleSkillActiveCommand(long SkillId);
